requirements.txt
click==6.7
-e git+git@github.com:wgwz/flask.git@a43a439c7a839edc8c5ac7062acb694eceb99e55#egg=Flask
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
python-dotenv==0.7.1
Werkzeug==0.13
